refactor: config validate module
