+++ src/validation/validate-config.js
@@ -1,5 +1,5 @@
 import checkConfigMissing from '../helpers/check-config-missing.js';
-import configError from '../helpers/config-error.js';
+import checkConfig from '../helpers/check-config.js';
 import getOptionValue from '../helpers/get-options.js';
 
 const validateConfig = () => {
@@ -8,33 +8,7 @@ const validateConfig = () => {
 
   checkConfigMissing(configValue);
 
-  const configCharsArr = configValue && configValue.split('');
-
-  configCharsArr.forEach((char, index, arr) => {
-    if (char.match(/[CR]/)) {
-      const modesAllowed = [0, 1];
-      const modeValue = arr[index + 1];
-      const isModeValid = modesAllowed.some(
-        (mode) => mode == modeValue,
-      );
-
-      if (!isModeValid) {
-        throw new configError(
-          'Available modes are: C1, C0 or R1, R0',
-        );
-      }
-    }
-    if (char.match(/A/)) {
-      const modeValue = arr[index + 1];
-      if (modeValue && modeValue !== '-') {
-        throw new configError(
-          'Atbash cipher should have not any modes',
-        );
-      }
-    }
-    if (char.match(/[^ACR0-9-]/)) {
-      throw new configError('Available ciphers are: C, A or R');
-    }
-  });
+  checkConfig(configValue);
 };
+
 export default validateConfig;
